console: Suppress build warnings
This patch silences the following warning:
ocfs2module.c:179: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 4 has type ‘__u64’

This patch silences the build warning by casting the argument as uint64_t.

+++ ocfs2console/ocfs2interface/ocfs2module.c
@@ -176,7 +176,7 @@ dinode_repr (DInode *self)
 {
   char blkno[32];
 
-  snprintf (blkno, sizeof (blkno), "%"PRIu64, self->dinode.i_blkno);
+  snprintf (blkno, sizeof (blkno), "%"PRIu64, (uint64_t)self->dinode.i_blkno);
   return PyString_FromFormat ("<ocfs2.DInode %s on %s>", blkno,
 			      PyString_AS_STRING (self->fs_obj->device));
 }
